+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/DebugAccountAt.java
@@ -105,7 +105,7 @@ public class DebugAccountAt extends AbstractBlockParameterOrBlockHashMethod {
     }
 
     List<TransactionWithMetadata> transactions = block.get().getTransactions();
-    if (transactions.isEmpty() || txIndex < 0 || txIndex > block.get().getTransactions().size()) {
+    if (transactions.isEmpty() || txIndex < 0 || txIndex >= transactions.size()) {
       return new JsonRpcErrorResponse(
           requestContext.getRequest().getId(), RpcErrorType.INVALID_TRANSACTION_PARAMS);
     }